requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
google-cloud-bigquery==3.23.1
scikit-learn==1.5.0
numpy==1.26.4
redis==5.0.4
